# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Restaurant Management SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants:
    get:
      operationId: listRestaurants
      tags: [restaurants]
      parameters:
        - { name: tenantId, in: query, schema: { type: integer } }
      responses:
        "200":
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
    post:
      operationId: createRestaurant
      tags: [restaurants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantInput"
      responses:
        "201":
          description: Created restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
  /restaurants/{id}:
    get:
      operationId: getRestaurant
      tags: [restaurants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"
    patch:
      operationId: updateRestaurant
      tags: [restaurants]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RestaurantUpdate"
      responses:
        "200":
          description: Updated restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Restaurant"

  # ─── Branches ─────────────────────────────────────────────────
  /restaurants/{restaurantId}/branches:
    get:
      operationId: listBranches
      tags: [branches]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"

  # ─── Users ────────────────────────────────────────────────────
...
  /restaurants/{restaurantId}/tables:
    get:
      operationId: listFloorTables
      tags: [floor-tables]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: List of floor tables
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FloorTable"
    post:
      operationId: createFloorTable
      tags: [floor-tables]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorTableInput"
      responses:
        "201":
          description: Created table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloorTable"
  /restaurants/{restaurantId}/tables/{id}:
    patch:
      operationId: updateFloorTable
      tags: [floor-tables]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FloorTableUpdate"
      responses:
        "200":
          description: Updated table
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FloorTable"
    delete:
      operationId: deleteFloorTable
      tags: [floor-tables]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted
  /restaurants/{restaurantId}/tables/{id}/qr:
    get:
      operationId: getTableQr
      tags: [floor-tables]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: QR code data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeData"

  # ─── Reservations ─────────────────────────────────────────────
  /restaurants/{restaurantId}/reservations:
    get:
      operationId: listReservations
      tags: [reservations]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: date, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string } }
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Created reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /restaurants/{restaurantId}/reservations/{id}:
    patch:
      operationId: updateReservation
      tags: [reservations]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ─── Menus ────────────────────────────────────────────────────
...
  /restaurants/{restaurantId}/attendance:
    get:
      operationId: listAttendance
      tags: [staff]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: userId, in: query, schema: { type: integer } }
        - { name: date, in: query, schema: { type: string } }
      responses:
        "200":
          description: Attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: clockIn
      tags: [staff]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockInInput"
      responses:
        "201":
          description: Clock in record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /restaurants/{restaurantId}/attendance/{id}/clock-out:
    patch:
      operationId: clockOut
      tags: [staff]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutInput"
      responses:
        "200":
          description: Clock out record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"
  /restaurants/{restaurantId}/audit-logs:
    get:
      operationId: listAuditLogs
      tags: [staff]
      parameters:
        - { name: restaurantId, in: path, required: true, schema: { type: integer } }
        - { name: userId, in: query, schema: { type: integer } }
        - { name: action, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: limit, in: query, schema: { type: integer } }
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # ─── Customers ────────────────────────────────────────────────
...
components:
  schemas:
...
    Restaurant:
      type: object
      required: [id, tenantId, name, slug]
      properties:
        id: { type: integer }
        tenantId: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        currency: { type: string }
        taxRate: { type: string }
        serviceCharge: { type: string }
        logoUrl: { type: ["string", "null"] }
        coverImageUrl: { type: ["string", "null"] }
        isActive: { type: boolean }
        openingTime: { type: string }
        closingTime: { type: string }
        createdAt: { type: string }
    RestaurantInput:
      type: object
      required: [tenantId, name, slug]
      properties:
        tenantId: { type: integer }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        city: { type: string }
        country: { type: string }
        taxRate: { type: string }
        openingTime: { type: string }
        closingTime: { type: string }
    RestaurantUpdate:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        phone: { type: string }
        email: { type: string }
        address: { type: string }
        city: { type: string }
        taxRate: { type: string }
        serviceCharge: { type: string }
        logoUrl: { type: string }
        openingTime: { type: string }
        closingTime: { type: string }

    # ── Branches ──
    Branch:
      type: object
      required: [id, restaurantId, name]
      properties:
        id: { type: integer }
        restaurantId: { type: integer }
        name: { type: string }
        address: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isMain: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string }
    BranchInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        address: { type: string }
        phone: { type: string }
        isMain: { type: boolean }

    # ── Users ──
...
    FloorTable:
      type: object
      required: [id, restaurantId, tableNumber, capacity, status]
      properties:
        id: { type: integer }
        restaurantId: { type: integer }
        tableNumber: { type: string }
        capacity: { type: integer }
        status: { type: string }
        positionX: { type: integer }
        positionY: { type: integer }
        shape: { type: string }
        qrCode: { type: ["string", "null"] }
        isActive: { type: boolean }
    FloorTableInput:
      type: object
      required: [tableNumber, capacity]
      properties:
        tableNumber: { type: string }
        capacity: { type: integer }
        positionX: { type: integer }
        positionY: { type: integer }
        shape: { type: string }
    FloorTableUpdate:
      type: object
      properties:
        tableNumber: { type: string }
        capacity: { type: integer }
        status: { type: string }
        positionX: { type: integer }
        positionY: { type: integer }
        shape: { type: string }
        isActive: { type: boolean }
    QrCodeData:
      type: object
      required: [qrUrl, tableNumber]
      properties:
        qrUrl: { type: string }
        tableNumber: { type: string }
        svgData: { type: string }

    # ── Reservations ──
    Reservation:
      type: object
      required: [id, restaurantId, guestName, scheduledAt, status]
      properties:
        id: { type: integer }
        restaurantId: { type: integer }
        tableId: { type: ["integer", "null"] }
        guestName: { type: string }
        guestPhone: { type: ["string", "null"] }
        guestEmail: { type: ["string", "null"] }
        partySize: { type: integer }
        scheduledAt: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    ReservationInput:
      type: object
      required: [guestName, scheduledAt, partySize]
      properties:
        guestName: { type: string }
        guestPhone: { type: string }
        guestEmail: { type: string }
        tableId: { type: integer }
        partySize: { type: integer }
        scheduledAt: { type: string }
        notes: { type: string }
    ReservationUpdate:
      type: object
      properties:
        guestName: { type: string }
        guestPhone: { type: string }
        tableId: { type: integer }
        partySize: { type: integer }
        scheduledAt: { type: string }
        status: { type: string }
        notes: { type: string }

    # ── Menus ──
...
    AttendanceRecord:
      type: object
      required: [id, userId, restaurantId, clockIn]
      properties:
        id: { type: integer }
        userId: { type: integer }
        restaurantId: { type: integer }
        clockIn: { type: string }
        clockOut: { type: ["string", "null"] }
        totalHours: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    ClockInInput:
      type: object
      required: [userId]
      properties:
        userId: { type: integer }
        notes: { type: string }
    ClockOutInput:
      type: object
      properties:
        notes: { type: string }
    AuditLog:
      type: object
      required: [id, action, entity, createdAt]
      properties:
        id: { type: integer }
        userId: { type: ["integer", "null"] }
        action: { type: string }
        entity: { type: string }
        entityId: { type: ["integer", "null"] }
        details: { type: ["string", "null"] }
        createdAt: { type: string }

    # ── Customers ──
...